+++ apps/admin/.eslintrc.js
@@ -2,7 +2,7 @@ module.exports = {
   root: true,
   extends: ["@repo/eslint-config/nextjs.js"],
   parserOptions: {
-    project: "./tsconfig.json",
+    project: ["./tsconfig.json", "../../packages/shared/tsconfig.json", "../../packages/database/tsconfig.json"],
     tsconfigRootDir: __dirname,
   },
 };